requirements.txt
asgiref==3.2.9
astroid==2.4.2
attrs==19.3.0
Babel==2.8.0
blis==0.4.1
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
click==4.1
cmake==3.18.0
colorama==0.4.3
cssselect2==0.3.0
cycler==0.10.0
cymem==2.0.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-appconf==1.0.4
django-cleanup==5.0.0
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-formtools==2.2
django-googledrive-storage==1.5.1
django-heroku==0.3.1
django-jsonfield==1.4.0
django-sendfile==0.3.11
django-storages==1.9.1
djangorestframework==3.12.2
docutils==0.15.2
docx2txt==0.8
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
html5lib==1.1
httplib2==0.19.0
idna==2.9
imutils==0.5.3
isort==4.3.21
jmespath==0.10.0
joblib==0.15.1
jsonschema==3.2.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
libmagic==1.0
mccabe==0.6.1
murmurhash==1.0.2
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.0.5
pdf2image==1.14.0
pdfminer.six==20200517
phonenumbers==8.12.6
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.9.7
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.9.5
pyrsistent==0.16.0
python-dateutil==2.8.1
python-magic==0.4.18
python3-openid==3.2.0
pytz==2020.1
regex==2020.6.8
reportlab==3.5.45
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
selenium==3.13.0
six==1.15.0
sortedcontainers==2.2.2
soupsieve==2.1
sqlparse==0.3.1
stripe==2.50.0
tinycss2==1.0.2
toml==0.10.1
tqdm==4.46.1
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.9
Wand==0.6.5
wasabi==0.6.0
webencodings==0.5.1
whitenoise==5.1.0
wrapt==1.12.1
xhtml2pdf==0.2.4
XlsxWriter==1.2.9
xmltodict==0.12.0